+++ tests/conftest.py
@@ -25,7 +25,7 @@ from .emulators import (  # noqa: F401
 )
 
 RAVEN_TESTING_DATA_BRANCH = os.getenv("RAVEN_TESTING_DATA_BRANCH", "master")
-SKIP_TEST_DATA = os.getenv("RAVENPY_SKIP_TEST_DATA")
+SKIP_TEST_DATA = True  # os.getenv("RAVENPY_SKIP_TEST_DATA")
 DEFAULT_CACHE = Path(_default_cache_dir)