+++ CHANGELOG.rst
@@ -5,6 +5,10 @@ Changelog
 v0.18.3 (unreleased)
 --------------------
 
+New features
+^^^^^^^^^^^^
+* Added `parsers.parse_rv` to extract a Command value from an RV file.
+
 Bug fixes
 ^^^^^^^^^
 * Fixed bug affecting `GriddedForcing`, where `station_idx` in the call to ``nc_specs`` was set to ``1`` instead of ``None``. (PR #501)